+++ src/ONTraC/data.py
@@ -1,12 +1,11 @@
 from optparse import Values
-from typing import Dict, List, Tuple
+from typing import Dict, List
 
 import numpy as np
 import pandas as pd
 import torch
 import torch_geometric.transforms as T
 from torch_geometric.data import Data, InMemoryDataset
-from torch_geometric.loader import DenseDataLoader
 
 from .log import *
 from .utils import count_lines, get_rel_params, read_yaml_file
@@ -22,7 +21,7 @@ class SpatailOmicsDataset(InMemoryDataset):
         super(SpatailOmicsDataset, self).__init__(root, transform, pre_transform)
 
     @property
-    def raw_file_names(self):
+    def raw_file_names(self):  # required by InMemoryDataset
         # return list(
         #     flatten([[sample for name, sample in data.items() if name != 'Name'] for data in self.params['Data']]))
         return []
@@ -46,8 +45,7 @@ class SpatailOmicsDataset(InMemoryDataset):
                 pos=torch.from_numpy(pd.read_csv(sample['Coordinates'])[['x', 'y']].values),
                 name=sample['Name'])
             data_list.append(data)
-        data, slices = self.collate(data_list)
-        torch.save((data, slices), self.processed_paths[0])
+        self.data, self.slices = self.collate(data_list)
 
 
 # ------------------------------------
@@ -65,13 +63,16 @@ def max_nodes(samples: List[Dict[str, str]]) -> int:
     return max_nodes
 
 
-def load_dataset(options: Values) -> Tuple[SpatailOmicsDataset, Data]:
+def load_dataset(options: Values) -> SpatailOmicsDataset:
+    """
+    Load dataset
+    :param options: Values, input options
+    :return: SpatailOmicsDataset, torch dataset
+    """
     params = read_yaml_file(f'{options.preprocessing_dir}/samples.yaml')
     rel_params = get_rel_params(options, params)
     dataset = create_torch_dataset(options, rel_params)
-    all_sample_loader = DenseDataLoader(dataset, batch_size=len(dataset))
-    data = next(iter(all_sample_loader))
-    return dataset, data
+    return dataset
 
 
 # ------------------------------------
@@ -84,17 +85,14 @@ def create_torch_dataset(options: Values, params: Dict) -> SpatailOmicsDataset:
     :return: None
     """
 
-    # ------------------------------------
     # Step 1: Get the maximum number of nodes
     m_nodes = max_nodes(params['Data'])
     # upcelling m_nodes to the nearest 100
     m_nodes = int(np.ceil(m_nodes / 100.0)) * 100
     info(f'Maximum number of cell in one sample is: {m_nodes}.')
-    # ------------------------------------
 
-    # ------------------------------------
     # Step 2: Create torch dataset
     dataset = SpatailOmicsDataset(root=options.preprocessing_dir, params=params,
                                   transform=T.ToDense(m_nodes))  # transform edge_index to adj matrix
-    # dataset = SpatailOmicsDataset(root=options.input, params=params)
+    dataset.process()
     return dataset